requirements.txt
fastapi
uvicorn
rembg
pillow
python-multipart  # per gestire upload file

